lc-sig: Make Pattern.t abstract, use internal signature
The signature `Pattern_S` duplicates the signature in the .mli file,
with the exception that it also declares the functions that are used
internally by LocalCompiler.

One observation is that each module has some analogue in SDN_Types. In
order to make these the types that they export properly abstract, the
modules themselves must handle the mapping to and from. For Pattern,
this has two consequences:

  * it must expose some way of making a 'singleton' pattern. This is the
    new `test_pat` function
  * it must expose some way of turning a Pattern.t into a
    SDN_Types.pattern. This is the `to_sdn` function, which a modified
    version of the `to_pattern` function that was previously in the
    Runtime module.