lc-sig: Make Action use an internal signature
The signature `Action_S` duplicates the signature in the .mli file,
with the exception that it also declares the functions that are used
internally by LocalCompiler.

Similar to Pattern, there are functions in in Runtime that translate an
Action.t to an 'SDN type'. But unlike Pattern.t, Action.t cannot be made
abstract in a straightforward way due to the `seq_act_pat` in the
Pattern module. In order to make Action properly abstract, the module
must specify a header, value, and pattern type that both Action and
Pattern can use to interoperate.