reactive: Fix definition of pick group
`ports` is a set of all all ports that are up on the switch. Given that,
the valid groups are the ones that only make reference to up ports.

This is a simplified view though. Since groups are parallel compositions
of sequenced actions, you could eliminate those parallel branches that
make refrence to not up ports. But keeping track of the combinatorics
for the purposes of efficient, incremental removal of flowtable rules
when a port goes down isn't feasible in the general.

In the common case, this may not be a problem.