reactive: Use proper liveness check for ports
The config component of the port description specifices whether or not
the port is installed on the switch. The state component of the port
description specifies whether or not a link is present at that part.
If config.down is true, then the port is not live. If config.down is
false, then the port is live if config.down is false.