query: initial implementation of query API
Add support to app API for defining queries. As of now, a query is an
identifier, an optional switch id, and a NetKAT predicate. Once the
controller recives these queries, it will run them periodically, which
will result in Query events that it will send back down to the
application.

Note that as of this commit, queries are not being executed on switches.
Applications registering queries will therefore not receive Query
events.