query: fix bugs in Transaction
First, do not start xids at 0, as that is the xid that's used when the
xid is irrelevant. This also allows the Transaction stage to skip over
messages with xid = 0 and avoid one-to-two Hashtbl lookups per message.

Second, send was generating two xids. The first was used to insert an
Ivar.t into the lookup table, and the second was the one that was
actually sent to the switch. This is no longer the case. send now
consistently uses a single xid.