discovery: add command-line flag
This commit breaks the controller in several ways when toplogy discovery
is enabled. Briefly, they are:

  * The controller crashes on switch disconnects; and
  * Consistent updates no longer work.

Both are due to the fact that the discovery application is what updates
the NIB when switches are added and removed. The code as written assumes
that if a switch disconnects, it's already in the NIB and therefore safe
to look up in the NIB. In the case of the controller processing
translating low-level openflow events to high-level network events, this
will cause a crash. In the case of the consistent update code, this
causes a silent failure. To fix this, responsibility for switch
bookkeeping should be shifted from the discovery application to the
controller itself.

Note that in addition, consistent updates will not actually provide
consistent updates when topology discovery is disabled. The code will
treat all ports as edge ports, which means that VLAN tags will never
survive a hop.

This is somewhat related to frenetic-lang/ocaml-topology#51.